requirements_firebase.txt
functions-framework==3.*
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
openai==1.10.0
Werkzeug==3.0.1
markdown2==2.4.12
reportlab==4.0.9
Pillow==10.2.0
pdfkit==1.0.0
